repo: Format source, organize imports and clean up
Bring the source code into the settings state by formatting the code,
organizing the imports, and running cleanup actions.

The code had random whitespace and imports out of order which cluttered
up pull requests for reviewing. We should keep our code well formatted
according to the formatting rules, etc. in the .settings folders. This
will simplify the changes in pull requests to make for easier reviewing
and merging.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>