Major refactoring
We separate the CLI aspects out of the transformer project into the cli
project. We use a TransformOptions interface as a callback from the
transformer library to the driver which can be the cli or the maven
plugin. Thus the maven plugin does not need to construct a command line
for the cli to parse.

We also separate the jakarta rules into a separate artifact which can
be used independently of the main lib or cli.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>