verboseness: Avoid leaking verbose/quiet abstraction into lib code
We now use --verbose or --quiet in the CLI to configure the log level
of the Logger passed to Transformer. For --verbose , the CLI will
configure the log level DEBUG. For --quiet, the CLI will configure the
log level ERROR.

In the lib code, we simply use debug level for verbose output, info for
normal output which wont be displayed if --quiet is specified to the
CLI which sets the log level to ERROR.

This then means we do not leak the CLI's verbose and quiet abstractions
into the lib code. Users of the lib code are free to configure log
levels as they see fit with out also needing to configure the verbose or
quiet app options.

We also use the Logger API directly in the code instead of having our
own "wrapper" methods. This allows for more efficient calling of logging
since we can avoid Object[] in many cases since we don't need to use the
varargs methods. We also pass any exception to the logger as excess
arguments rather than having our own special handling code. This allows
the logger implementation to handle exceptions in its own way.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>