refactor: Hoist ByteData up into action API
We use an interface for ByteData to hold a named ByteBuffer. ByteBuffer
already encapsulates a byte array, offset, and length and is a type
used with the input and output stream classes in IO.

We also use a ThreadLocal to hold the reused ByteBuffer so multiple
threads can use the transformer API.

Finally, we add a method to Action which takes a ByteData as input as
this is used by CompositeActionImpl and we need to support Action
implementations which do not inherit from ActionImpl.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>