renaming: Better handle package renaming for non-class resources
We refactor NullActionImpl to be RenameActionImpl since even when
not transforming contents, it does need to handle package renames
to place the resources of a package in the proper place when the
package is renamed. Otherwise the resources are orphaned in the
original package name.

We also update DirectoryActionImpl and ZipActionImpl to handle the
necessary renaming when using the stream and file apply methods.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>