transformer: Refactor to use Map instead of Properties in many places
This makes other operations easier and we want Maps in the end anyway.

loadSubstitution is updated to handle a comma-separated value where
each item in the list is read as a properties file and merged in.
So master text and per-direct entries can specify multiple properties
files. This has the side effect of fixing per-direct support.

Fixes https://github.com/eclipse/transformer/issues/185

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>