encoding: Resource encoding can be specified in selection rules
This change uses the values of selection rules to specify the charset
name for the matched resource names. Previously the selection rule
value was ignored.

In another change from before, to specify a selection exclusion, the
value (charset name) must start with '!'. This change is necessary since
the prior behavior of prefixing the selection rule key with '!' does
not work in properties files since lines starting with '!' are comments.

The default selection rule is "*=UTF-8" which selects all resource names
and uses the charset UTF-8.

Fixes https://github.com/eclipse/transformer/issues/293

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>