New line codes in text files are always converted to LF only
We implement a BufferedReader subclass to enable us to obtain the
actual line ending for each read line so that we can use it when
writing out the transformed line.

Since we need to use this BufferedReader subclass, we change ByteData
to return an unbuffered Reader so we don't end up double buffering.

Fixes #292

Signed-off-by: Ryosuke Okada <ryosuke.okada@fujitsu.com>
Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>